Refactor Send screen state and separate nearby device UI data
Moves the Send screen toward a clearer MVVM/UDF structure.

The Send screen now reads one SendScreenState from SendScreenViewModel instead of keeping important screen state inside the composable. The ViewModel now owns the selected tab, text input, text send status, nearby devices, and discovery status.

This makes the Send screen easier to reason about:

state flows down
events go up
ViewModel owns the screen state
Composable renders the state

Added TextSendState for the text sending lifecycle:

- Idle
- Sending
- Sent
- Failed

This replaces the older generic send status shape and makes it clear that the current send status belongs to text transfer.

The Send screen now sends events to the ViewModel for:

- text changes
- tab changes
- clearing text send status
- restarting discovery
- sending text to a selected device

Removed the old SendItem-based text flow.

Text is no longer treated like a selected item list. The current text input is now the text that will be sent. This keeps text sending simple and leaves selected-list behavior for future file sending, where it actually makes sense.

Renamed the old send item card into a text-specific card and removed the unused SendItem model.

Separated nearby device UI data from the domain model.

Network discovery still produces NearbyDevice domain models, because those contain the real data needed for sending, such as host addresses and port. The ViewModel now keeps the latest domain devices privately and maps them into NearbyDeviceUiModel for the Send screen.

The UI renders NearbyDeviceUiModel and sends only the device id back when the user taps a device. The ViewModel then looks up the real NearbyDevice by id before sending text.

This keeps the direction clear:

NearbyDevice = domain/network data
NearbyDeviceUiModel = UI rendering data
device id = event data from UI back to ViewModel

Also cleaned the Android 13 and lower NSD resolve path.

The old resolveService path had duplicate nearby-device updates. It now resolves the service, converts it to NearbyDevice, skips the local device, and adds or replaces the device in the list by id.

That means if the same device is resolved again with updated host or port info, the old entry is replaced with the new one.

No file transfer behavior was added here. This is a cleanup step for Send screen state, text sending UI state, nearby device UI mapping, and Android discovery resolve logic.
