--- shared/src/commonMain/kotlin/com/liftley/sync360/presentation/featureSend/model/SendItem.kt
@@ -1,5 +1,0 @@
-package com.liftley.sync360.presentation.featureSend.model
-
-interface SendItem {
-    data class Text(val text: String): SendItem
-}
